Updatedet DeleteObjects

+++ src/main/java/dk/kb/cop3/backend/scripts/DeleteObjects.java
@@ -3,6 +3,7 @@ package dk.kb.cop3.backend.scripts;
 import dk.kb.cop3.backend.constants.CopBackendProperties;
 import dk.kb.cop3.backend.crud.database.HibernateUtil;
 import dk.kb.cop3.backend.crud.database.hibernate.Category;
+import dk.kb.cop3.backend.crud.database.hibernate.Edition;
 import dk.kb.cop3.backend.crud.database.hibernate.Object;
 import dk.kb.cop3.backend.migrate.TestSearch;
 import org.apache.solr.client.solrj.SolrServerException;
@@ -26,26 +27,41 @@ public class DeleteObjects {
     public static void main(String[] args) throws SolrServerException, IOException {
         String mode = args[0];
 
-        initialize();
-        if ("object".equals(mode)) {
-            deleteSingleObject(args[1]);
-            System.exit(0);
+        if (args.length == 2 ) {
+            initialize();
+            if ("object".equals(mode)) {
+                deleteSingleObject(args[1]);
+                System.exit(0);
+            }
+            if ("category".equals(mode)) {
+                deleteAllObjectsInCategory(args[1]);
+                System.exit(0);
+            }
+            if ("edition".equals(mode)) {
+                deleteEdition(args[1]);
+                System.exit(0);
+            }
         }
-        if ("category".equals(mode)) {
-            deleteAllObjectsInCategory(args[1]);
-            System.exit(0);
-        }
-        if ("edition".equals(mode)) {
-            deleteEdition(args[1]);
-            System.exit(0);
-        }
+        System.out.println("Usage:");
+        System.out.println("Delete object:");
+        System.out.println("java -Ddk.kb.cop.propertiesFile=/path/to/cop-config.xml -cp cop3-backend-cop-jar.jar dk.kb.cop3.backend.scripts.DeleteObjects object objectID");
+        System.out.println("Delete category:");
+        System.out.println("java -Ddk.kb.cop.propertiesFile=/path/to/cop-config.xml -cp cop3-backend-cop-jar.jar dk.kb.cop3.backend.scripts.DeleteObjects category categoryID");
+        System.out.println("Delete edition:");
+        System.out.println("java -Ddk.kb.cop.propertiesFile=/path/to/cop-config.xml -cp cop3-backend-cop-jar.jar dk.kb.cop3.backend.scripts.DeleteObjects edition editionID");
     }
 
     private static void deleteEdition(String arg) {
+        Edition edition = session.get(Edition.class,arg);
+        if (edition != null) {
+            List<Object> objectsToBeDeleted = session.createQuery("select o from dk.kb.cop3.backend.crud.database.hibernate.Object o where o.edition.id = '"+arg+"'").list();
+            objectsToBeDeleted.stream().forEach(DeleteObjects::deleteObject);
+        } else {
+            System.out.println("Edition not found "+arg);
+        }
     }
 
     private static void deleteAllObjectsInCategory(String arg) {
-        System.out.println("category "+arg);
         Category category = session.get(Category.class,arg+"/da/");
         if (category != null) {
             List<Object> objectsToBeDeleted = session.createQuery("select o from dk.kb.cop3.backend.crud.database.hibernate.Object o join o.categories c where c.id = '"+arg+"/da/'").list();
@@ -60,7 +76,7 @@ public class DeleteObjects {
         if (objectToBeDeleted != null) {
             deleteObject(objectToBeDeleted);
         } else {
-            System.out.println("object not found");
+            System.out.println("No object found with id " + objectID);
         }
     }
 
@@ -72,10 +88,8 @@ public class DeleteObjects {
             solr.commit();
             session.delete(objectToBeDeleted);
             transaction.commit();
-        } catch (SolrServerException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        } catch (SolrServerException | IOException e) {
+            System.out.println("Error deleting object "+objectToBeDeleted.getId()+": "+e.getMessage());
         }
     }
